requirements.txt
matplotlib==3.5.3
numpy==1.22.4
pandas==1.5.2
Pillow==9.3.0
Pillow==10.3.0
python_dateutil==2.8.2
streamlit==1.35.0
